First Commit For Day 2

+++ Day 1/Calculator.java
@@ -1,30 +1,30 @@
 class MathOperator{
-	public int Addition(int A, int B){
-		int Result;
+	public float Addition(float A, float B){
+		float Result;
 		Result = A + B;
 		return Result;
 	}
 
-	public int Subtraction(int A, int B){
-		int Result;
+	public float Subtraction(float A, float B){
+		float Result;
 		Result = A - B;
 		return Result;
 	}
 
-	public int Multiplication(int A, int B){
-		int Result;
+	public float Multiplication(float A, float B){
+		float Result;
 		Result = A * B;
 		return Result;
 	}
 
-	public int Division(int A, int B){
-		int Result;
+	public float Division(float A, float B){
+		float Result;
 		Result = A / B;
 		return Result;
 	}
 
-	public int Modulation(int A, int B){
-		int Result;
+	public float Modulation(float A, float B){
+		float Result;
 		Result = A % B;
 		return Result;
 	}
@@ -34,10 +34,10 @@ class Calculator{
 	public static void main(String XYZ[]){
 		MathOperator MOp;
 		MOp = new MathOperator();
-		System.out.println("Result A: " + MOp.Addition(50,32));
-		System.out.println("Result B: " + MOp.Subtraction(50,32));
-		System.out.println("Result C: " + MOp.Multiplication(50,32));
-		System.out.println("Result D: " + MOp.Division(50,32));
-		System.out.println("Result E: " + MOp.Modulation(50,32));
+		System.out.println("Result A: " + MOp.Addition(50.0f,32.0f));
+		System.out.println("Result B: " + MOp.Subtraction(50.0f,32.0f));
+		System.out.println("Result C: " + MOp.Multiplication(50.0f,32.0f));
+		System.out.println("Result D: " + MOp.Division(50.0f,32.0f));
+		System.out.println("Result E: " + MOp.Modulation(50.0f,32.0f));
 	}
 }
